Move static class from HugeGF to own top-level classes
