+++ core/src/main/java/org/neo4j/graphalgo/core/huge/RelationshipLoader.java
@@ -1,0 +1,8 @@
+package org.neo4j.graphalgo.core.huge;
+
+import org.neo4j.kernel.api.exceptions.EntityNotFoundException;
+
+@FunctionalInterface
+interface RelationshipLoader {
+    void apply(long neoId, long nodeId) throws EntityNotFoundException;
+}
